docs/requirements.txt
mkdocs==1.6.1
mkdocs-material==9.5.39
mkdocs-minify-plugin==0.8.0
mike==2.1.3
pymdown-extensions==10.14.2
PyYAML==6.0.2
